+++ src/main/java/com/patiun/libraryspring/configuration/MvcConfig.java
@@ -14,6 +14,7 @@ import java.util.Locale;
 @Configuration
 public class MvcConfig implements WebMvcConfigurer {
 
+    @Override
     public void addViewControllers(ViewControllerRegistry registry) {
         registry.addViewController("/").setViewName("home");
         registry.addViewController("/error").setViewName("error");
