+++ src/main/java/com/patiun/libraryspring/order/BookOrderRepository.java
@@ -3,6 +3,10 @@ package com.patiun.libraryspring.order;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface BookOrderRepository extends CrudRepository<BookOrder, Integer> {
+
+    List<BookOrder> findByUserId(Integer id);
 }
